+++ main.rb
@@ -1,4 +1,6 @@
 require './app'
+require './handle_choice'
+require './main_fun'
 
 def options
   puts ''
@@ -10,47 +12,6 @@ def options
   puts '5 - Create a rental.'
   puts '6 - List all rentals for a given person id.'
   puts '7 - Exit.'
-end
-
-def handle_choice(choice, app)
-  case choice
-
-  when 3
-    app.create_person
-
-  when 4
-    app.create_book
-
-  when 5
-    app.create_rental
-
-  when 6
-    app.rental_list
-
-  when 7
-    puts ':) Thank you for using this app!'
-    return false
-
-  end
-  true
-end
-
-def main
-  app = App.new
-  puts 'Welcome to OOP School Library App!'
-  proceed = true
-
-  while proceed
-    options
-    choice = gets.chomp.to_i
-    if choice == 1
-      app.book_list
-    elsif choice == 2
-      app.people_list
-    else
-      proceed = handle_choice(choice, app)
-    end
-  end
 end
 
 main
